moved more node functions to helpers file
